Update bitmap.hx to allow set_tile to be public
The intent is to allow the tile bitmap to be updated without destroying and re-creating the object.  In my specific example I wanted to include the bitmap in a flow but didn't want to destroy the old bitmap  and instance a new one for fear that it would change my flow layout.   I added the tile= null line assuming it would assist in garbage collection.